+++ Django_REST_API/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for delivery_hero project.
+Django settings for django_rest_api project.
 
 Generated by 'django-admin startproject' using Django 2.0.4.
 
@@ -53,7 +53,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'delivery_hero.urls'
+ROOT_URLCONF = 'Django_REST_API.urls'
 
 # Swagger settings ensuring that swagger shows token based authentication pop up.
 SWAGGER_SETTINGS = {
@@ -84,7 +84,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'delivery_hero.wsgi.application'
+WSGI_APPLICATION = 'Django_REST_API.wsgi.application'
 
 
 # Database
