+++ Django_REST_API/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for delivery_hero project.
+WSGI config for django_rest_api project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "delivery_hero.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "Django_REST_API.settings")
 
 application = get_wsgi_application()
